+++ tests/abi/c_smoke.c
@@ -30,7 +30,7 @@ int main(int argc, char **argv) {
     return 64;
   }
   (void)remove(argv[1]);
-  if (pp_abi_version() != UINT32_C(21)) {
+  if (pp_abi_version() != UINT32_C(22)) {
     return 1;
   }
   pp_error_code_t status =